smartcard/python/requirements.txt
cryptography==40.0.1
petlib==0.0.45
pycryptodomex==3.17
